requirements.txt
openai
gspread
google-auth
pinecone-client
cassandra-driver
langchain
tiktoken
protobuf~=3.19.0
